requirements.txt
flask==2.3.3
python-dotenv==1.0.1
pytest==7.4.0
flask-cors==4.0.0
gunicorn==21.2.0 
